hackathon_2_phase_3/backend/pyproject.toml
[project]
name = "hackathon-todo-backend"
version = "1.0.0"
description = "FastAPI backend for Hackathon II Todo app"
requires-python = ">=3.13"

dependencies = [
    # FastAPI & ASGI Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Database & ORM
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.0",

    # Async PostgreSQL Driver
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",

    # Authentication
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.17",
    "python-jose[cryptography]>=3.3.0",

    # Environment & Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    # Validation & Serialization
    "email-validator>=2.1.0",

    # Testing (dev)
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
